pyproject.toml
[project]
name = "ideathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "groq>=1.1.1",
    "openai>=2.29.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "pyttsx3>=2.99",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "tavily>=1.1.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
]
